fix code review point 3 (strange paths and repository combinations) and null pointer exception when adding a relative path without parent

+++ src/darep/Helper.java
@@ -51,6 +51,9 @@ public class Helper {
 	}
 	
 	public static boolean deleteDir(File dir) {
+		if (!dir.exists()) {
+			return true;
+		}
 		if (dir.isDirectory()) {
 			for (File subfile: dir.listFiles()) {
 				boolean success = deleteDir(subfile);
